Added a new DbHelper object to be used in the ado.net context and schema config helper for accessing the database
